backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
prophet==1.1.4
statsmodels==0.14.0
pandas==2.0.3
numpy==1.25.2
pydantic==2.3.0
python-dotenv==1.0.0
httpx==0.24.1
supabase==1.0.3 